client/requirements.txt
fastavro
mysql-connector-python
pandas
numpy

